build: allow falcoctl to be built on non-linux machines by disabling the probe loader on GOOS != linux
